[actuator] remove the generic in Actuator
The generic has been removed because Number is not comparable, therefore a type would have been required to be chosen in the implementation, since the position has to be limited.
Avoiding the use of floating-point types should give better control of the discrete positions.
int has been chosen because a 32-bit resolution should be well enough for an encoder in a domotic actuator.