[window] refactor MechanizedWindow with Actuator in the constructor
This allows the creation of windows with different kinds of actuators:
it might be desired to use a simple actuator with an instantaneous rate of change,
or add one with a limited rate of movement, or even make it even more complex
with nth order lags.
By passing the actuator in the constructor it is possible to do this without having
to create other almost identical window implementations.